Removed unnecessary BatchFilterModel class
